Fixed imports, moved things from server to main, added run commands
