+++ .isort.cfg
@@ -1,2 +1,2 @@
-[tool.isort]
-profile = "black"
+[settings]
+profile = black
