+++ src/ephys_link/platforms/new_scale_handler.py
@@ -13,8 +13,8 @@ import clr
 # noinspection PyPackageRequirements
 import socketio
 
-from src.ephys_link import common as com
-from src.ephys_link.platform_handler import PlatformHandler
+from ephys_link import common as com
+from ephys_link.platform_handler import PlatformHandler
 
 
 class NewScaleHandler(PlatformHandler):
@@ -28,7 +28,7 @@ class NewScaleHandler(PlatformHandler):
 
         # Load New Scale API
         # noinspection PyUnresolvedReferences
-        clr.AddReference("../resources/NstMotorCtrl")
+        clr.AddReference("ephys_link/resources/NstMotorCtrl")
         # noinspection PyUnresolvedReferences
         from NstMotorCtrl import NstCtrlHostIntf
