+++ src/ephys_link/platforms/sensapex_handler.py
@@ -12,8 +12,8 @@ import socketio
 from sensapex import UMP, UMError
 from sensapex_manipulator import SensapexManipulator
 
-import src.ephys_link.common as com
-from src.ephys_link.platform_handler import PlatformHandler
+import ephys_link.common as com
+from ephys_link.platform_handler import PlatformHandler
 
 
 class SensapexHandler(PlatformHandler):
